v1.2.5 - Fix multiple state detection issues
Issue:

Previously, the MQTT callback kept triggering a motor start event, which in turn triggered a motor start delay event. This cycle caused the start of the motor to be delayed until the printer status moved off of 14 (Cleaning Nozzle).

Solution:

The fix ensures that the motor start event is only triggered when the motor is not in a waiting or running state, preventing repeated delays.